refactor(vaultManager): move liquidation thresholds to liquidation.js
